avanzatech_blog/requirements.txt
Django==5.0
djangorestframework==3.14.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
